requirements.txt
django-extensions==1.6.7
django-regex-field==0.2.0
django-tables2==1.2.5
django-taggit==0.20.2
